pyproject.toml
[tool.poetry]
name = "chat-websockets"
version = "0.1.0"
description = ""
authors = ["Aleksandr <aleksandr.voitushenko@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
requests = "^2.32.3"
aiohttp = "^3.9.5"
python-dotenv = "^1.0.1"
websockets = "^12.0"
names = "^0.3.0"
aiofile = "^3.8.8"
aiopath = "0.6.11"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
